fix bug in atd generation
It seems previously we where using an old, deprecated version of of
the atdgen command line. This moves us to the correct generation
model. That is `atdgen -j -j-std` instead of just `atdgen -j-std`
which essentially has the same semantics as `atdgen -json`. This also
changes the output name from <module name> to <module name>_j and
<module name>_t. Overall this solves a class of internal generation
problems as well.